adapt framework_agreement_department to portfolios
