+++ framework_agreement_department/__openerp__.py
@@ -17,12 +17,13 @@
 
 {
     'name': 'Framework Agreements with Department Categorization',
-    'version': '1.0',
+    'version': '2.0',
     'category': 'Generic Modules/Sales & Purchases',
     'author': "Camptocamp,Odoo Community Association (OCA)",
     "license": "AGPL-3",
     'website': 'http://camptocamp.com',
     'depends': ['framework_agreement', 'hr'],
-    'data': ['view/agreement.xml'],
+    'data': ['view/agreement.xml',
+             'view/portfolio.xml'],
     'installable': True,
 }
