+++ framework_agreement_department/model/portfolio.py
@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 # Author: Leonardo Pistone
-# Copyright 2014 Camptocamp SA (http://www.camptocamp.com)
+# Copyright 2015 Camptocamp SA (http://www.camptocamp.com)
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -15,5 +14,16 @@
 # You should have received a copy of the GNU Affero General Public Lice
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from . import agreement
-from . import portfolio
+from openerp import models, fields
+
+
+class Portfolio(models.Model):
+    _inherit = 'framework.agreement.portfolio'
+
+    def _get_my_department(self):
+        employees = self.env.user.employee_ids
+        return (employees and employees[0].department_id or
+                self.env['hr.department'])
+
+    department_id = fields.Many2one('hr.department', 'Department',
+                                    default=_get_my_department)
